requirements.txt
aws-cdk-lib==2.23.0
constructs>=10.0.0,<11.0.0
pwed-cdk>=0.0.51
